Removed purpose template selection
